Move sprite-related files to sprites directory
Prior to this change, the sprite files were in the main rpg directory.

This change moves these files to a sprite directory.
