+++ rpg/load_game_map.py
@@ -9,9 +9,9 @@ from os.path import isfile, join
 import arcade
 from arcade.experimental.lights import Light, LightLayer
 
-from rpg.character_sprite import CharacterSprite
+from rpg.sprites.character_sprite import CharacterSprite
 from rpg.constants import TILE_SCALING
-from rpg.path_following_sprite import PathFollowingSprite
+from rpg.sprites.path_following_sprite import PathFollowingSprite
 
 
 class GameMap:
