+++ rpg/sprites/path_following_sprite.py
@@ -2,7 +2,7 @@ import math
 
 import arcade
 
-from rpg.character_sprite import CharacterSprite
+from rpg.sprites.character_sprite import CharacterSprite
 
 
 class PathFollowingSprite(CharacterSprite):
