+++ rpg/views/game_view.py
@@ -9,7 +9,7 @@ import rpg.constants as constants
 from arcade.experimental.lights import Light
 from pyglet.math import Vec2
 from rpg.message_box import MessageBox
-from rpg.player_sprite import PlayerSprite
+from rpg.sprites.player_sprite import PlayerSprite
 
 
 class GameView(arcade.View):
